Splittet some test-files into correct test-files
added file for tests on primitive types
added tests for txt-types
