+++ code/src/test/kotlin/parserTests/TupleTests.kt
@@ -10,7 +10,6 @@ import parser.ParserWithoutBuiltins
 
 class TupleTests
 {
-
     @org.junit.jupiter.api.Test
     fun parseTupleTypeWithoutAssignment() {
         assertThat(
@@ -25,9 +24,11 @@ class TupleTests
     fun parseTupleWithAssignment() {
         assertThat(
             buildTokenSequence {
-                tuple.squareStart.number.`,`.text.squareEnd.identifier("myTuple").`=`.`(`.number(5.0).`,`.text("someText").`)`.newLine
+                tuple.squareStart.number.`,`.text.squareEnd.identifier("myTuple").`=`.`(`.number(5.0)
+                        .`,`.text("someText").`)`.newLine
             },
-            matchesAstChildren("myTuple" declaredAs tpl(num, txt) withValue tpl(num(5), txt("someText")))
+            matchesAstChildren("myTuple" declaredAs tpl(num, txt) withValue tpl(num(5),
+                               txt("someText")))
         )
     }
 
@@ -44,17 +45,21 @@ class TupleTests
     @org.junit.jupiter.api.Test
     fun failsOnNotEnoughSeparators() {
         val lexer = DummyLexer(buildTokenSequence {
-            tuple.squareStart.number.`,`.text.squareEnd.identifier("myTuple").`=`.`(`.number(5.0).text("someText").`)`.newLine
+            tuple.squareStart.number.`,`.text.squareEnd.identifier("myTuple").`=`.`(`.number(5.0)
+                    .text("someText").`)`.newLine
         })
-        Assertions.assertThrows(WrongTokenTypeError::class.java) { ParserWithoutBuiltins(lexer).generateAbstractSyntaxTree() }
+        Assertions.assertThrows(WrongTokenTypeError::class.java)
+            { ParserWithoutBuiltins(lexer).generateAbstractSyntaxTree() }
     }
 
     @org.junit.jupiter.api.Test
     fun failsOnTupleAssignmentFailsOnTooManySeparators() {
         val lexer = DummyLexer(buildTokenSequence {
-            tuple.squareStart.number.`,`.text.squareEnd.identifier("myTuple").`=`.`(`.number(5.0).`,`.`,`.text("someText").`)`.newLine
+            tuple.squareStart.number.`,`.text.squareEnd.identifier("myTuple").`=`.`(`.number(5.0)
+                    .`,`.`,`.text("someText").`)`.newLine
         })
-        Assertions.assertThrows(UnexpectedTokenError::class.java) { ParserWithoutBuiltins(lexer).generateAbstractSyntaxTree() }
+        Assertions.assertThrows(UnexpectedTokenError::class.java)
+            { ParserWithoutBuiltins(lexer).generateAbstractSyntaxTree() }
     }
 
 }
